extract nested LambdaStack
