+++ packages/geoprocessing/scripts/aws/publicBuckets.ts
@@ -32,7 +32,7 @@ export const createPublicBuckets = (
     }),
   };
 
-  if (stack.functions.processingFunctions.length > 0) {
+  if (stack.lambdaStack.getProcessingFunctions.length > 0) {
     const result = new Bucket(stack, `GpResultBucket`, {
       bucketName: `gp-${stack.props.projectName}-results`,
       versioned: false,
